- Moved TclX_AppendObjResult to be an external API for testing - Upgraded the patch levels to be "8.1.2". - Defined TclX_MainEx to take an interp as an argument.  TclX_Main is   now a macro that calls TclX_MainEx.  When passing the interp   argument, Tcl_CreateInterp() is called.  This is done because the   Tcl API can only be called inside the "exe" and not in the "dll." - Added type cast to SetMatchInfoVar API. The bug appeared due to the new   level of warnings used while compiling TclX. - Added Tcl_InitStub call to TclX_Init API so that stubs are loaded   when the package is loaded into a running Tcl shell.

+++ tcl/generic/tclXfilescan.c
@@ -13,7 +13,7 @@
  * software for any purpose.  It is provided "as is" without express or
  * implied warranty.
  *-----------------------------------------------------------------------------
- * $Id$
+ * $Id: tclXfilescan.c,v 8.20 1999/04/09 00:21:39 markd Exp $
  *-----------------------------------------------------------------------------
  */
 
@@ -582,7 +582,7 @@ SetMatchInfoVar (interp, scanData)
     }
     regExpPtr = (TclRegexp *) scanData->matchPtr->regExp;
 
-    for (idx = 0; idx < regExpPtr->re.re_nsub; idx++) {
+    for (idx = 0; (unsigned int) idx < regExpPtr->re.re_nsub; idx++) {
 	start = regExpPtr->matches[idx+1].rm_so;
 	end = regExpPtr->matches[idx+1].rm_eo;
